+++ lib/engine/game/base.rb
@@ -1534,6 +1534,36 @@ module Engine
         ability.hexes.include?(hex.id)
       end
 
+      def rust_trains!(train, entity)
+        obsolete_trains = []
+        rusted_trains = []
+        owners = Hash.new(0)
+
+        trains.each do |t|
+          next if t.obsolete || t.obsolete_on != train.sym
+
+          obsolete_trains << t.name
+          t.obsolete = true
+        end
+
+        trains.each do |t|
+          next if t.rusted
+
+          should_rust = t.rusts_on == train.sym || (t.obsolete_on == train.sym && @depot.discarded.include?(t))
+          next unless should_rust
+          next unless rust?(t)
+
+          rusted_trains << t.name
+          owners[t.owner.name] += 1
+          entity.rusted_self = true if entity && entity == t.owner
+          rust(t)
+        end
+
+        @log << "-- Event: #{obsolete_trains.uniq.join(', ')} trains are obsolete --" if obsolete_trains.any?
+        @log << "-- Event: #{rusted_trains.uniq.join(', ')} trains rust " \
+          "( #{owners.map { |c, t| "#{c} x#{t}" }.join(', ')}) --" if rusted_trains.any?
+      end
+
       private
 
       def init_bank
